+++ index.html
@@ -9,6 +9,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="/js/functions.js"></script>
   <script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
